Create Q2PartyCategory.dart
Q2. (Asked by Amazon) O(N)
A professor went to a party. 
Being an erudite person, he classified 
the party into two categories. 
He proposed that if all the persons 
in the party are wearing different 
colored robes, then that is a girl’s 
only party. If we even get one 
duplicate color, it must be a boy’s 
party. The colors of the robes are 
represented by positive integers.

Input
The first line of each test case 
contains an integer T, denoting the 
no of test cases. Then T test cases 
follow. The first line of each 
test case contains an integer N denoting
the number of people in the party. 
In the next line are N space separated 
values of Ai denoting the color of the 
robes.

Output
For each test case, print “BOYS” 
without quotes if it’s a boy’s party, 
else print “GIRLS”